requirements.txt
cloudshell-core==2.2.176
cloudshell-L1-networking-core>=1.0,<1.1
cloudshell-cli>=3.3,<3.4
